Updated for go 1.
This mostly involved replacing []interface{} in many places with
datastore.Value

I modified conn_test.go somewhat because I think it was wanting the
testdb to be under role gopqtest db. The way it works now is it expects
the user invoking the test to be the owner (or at least have the
proper permissions) for gopgtest.